requirements.txt
pyyaml
pandas
scikit-learn
transformers
accelerate